+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ##### Lobby
 * NEW: Matchmaker now support allies. Go to maniaplanet live chat in order to set one of your buddy as ally
-* IMPROVEMENT: For backups player time before jumping has been increased
 * NEW: Player's rank is now displayed in the player list
+* NEW: Text is now different for ready players if there is enough players ready or not
+* NEW: Rename AlliesElite in Elite and former Elite MatchMaker is now known as DistanceElite
+* IMPROVEMENT: For backups player time before jumping has been increased
 
 # 0.1